Fixed apache::vhost. IMPORTANT: Read comment details
The existing apache::vhost define was almost not used in Example42 puppet modules and had a bug
that prevented it from working properly.
The new version of apache::vhost is going to be used by the new modules (instead of the less flexible apache::virtualhost)
such as the worpress one.

This change is therefore needed but makes a potentially dangerous change in the name of the files under Ubuntu that may break your Apache
if you happened to use it somewhere (suppose in a modified version, since the existing one was not working... )