Initial Import from example42 Puppet control-repo
This very first release of PSICK is based on example42 control-repo:
a lot of Puppet code and data for a 0.0.1 release ;-)

The repository https://github.com/example42/control-repo
is renamed to https://github.com/example42/psick
the whole commit history is zeroed.

The original repo with the full git history has been cloned to
https://github.com/example42/control-repo-archive
for reference but is not going to have furthher changes.

We did we do such a terrible thing?

  - This repo has never intended to be an upstream origin to sync with:
    each Puppet control-repo is intrincecally a non convergent fork

  - PSICK is based on the existing control-repo but has broader tasks
    and goals: it's a brand new prject starting from robust foundatiouns

  - We want the new project to start from a lean and clean repo, without
    the weight of 8 years or more or less unrelated code

  - We wanted to preserve GitHub's stars on a 8 years long idea which is
    still here: a single place from where to manage a whole infrastructure
    via Puppet.

How to cope with local forks of the original https://github.com/example42/control-repo:

Our suggestion is to simply clone https://github.com/example42/psick from scratch
and use is as source for additions and code for your own local control-repo.

PSICK purpose (still not fully implemented) is easily the users pick from PSICK repo
componenents or updates and then integrate them in their local, with a separated history, repo.